Update from version 0.2.0 to 0.3.0.
**Background**
This update addresses the issue related to API changes in the library. Specifically, the change involves the `Vector::distance` method, which has been replaced with the `Metric::measure` method to calculate the distance between vectors. The updated API now follows the signature:
```diff
- Vector::distance(&self, other: &Self, metric: Metric) -> T;
+ Metric::measure(&self, v0: &[T], v1: &[T]) -> T;
```

**Changes**
- Modified the `metric` module to provide distance functions and compare functions.
- Moved some utilities related to the `simd` feature to the public scope and documented them. This enhances maintainability.
- Added more various check routines to the CI file (`main.yml`) for improved quality assurance.
- Added additional documentation e.g. in `src/metric/metric.rs`, simple formulas are provided for every metric.

**Reasons**
The API change was made to improve clarity and consistency within the library. By introducing the `Metric::measure` method, it provides a more descriptive and intuitive way to calculate distances between vectors. Additionally, refactoring the `simd` module enhances maintainability, making it easier to maintain and extend the codebase.